refactor: extract model-provider contract to core, break workflow->agent cycle
model-profile (workflow layer) imported AGENT_MODEL_APIS + model types from
agent/host.ts, forming a real dependency cycle (agent->workflow is dominant at
18 edges; this reverse edge closed the loop).

Move the host-neutral model provider contract into src/core/model-provider.ts:
it is a provider capability contract, not agent-specific. host.ts re-exports for
agent-layer consumers; model-profile now imports from core.

Verified: tsc --noEmit clean, 486 vitest tests pass, tsc build clean.
